requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.2
django-colorfield==0.10.1
django-cors-headers==3.14.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
drf_extra_fields==3.4.0
drf-spectacular==0.25.1
filelock==3.11.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.3
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.1.0
identify==2.5.22
idna==3.4
importlib-metadata==1.7.0
importlib-resources==5.12.0
inflection==0.5.1
isort==5.11.5
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mccabe==0.7.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.1
pep8-naming==0.13.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pre-commit==3.2.2
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
zipp==3.15.0
